Add experimental 'packages' corral.json entry
The packages entry lays out which packages this bundle provides.
'packages' will be used as when putting together a release for
a given bundle. Only the packages provided will be included in
the release artifact.

It will also allow tool like documentation generation to figure out
where the various packages it is using are coming from.